Split scene validation
